Fixed problem with connecting to database
I had a problem connecting to the original database that you had given me so I was unable to work with the posts in the database. Up until this point, I was working with my own created database tables and posts so I wouldn't be hung up on this particular issue. I solved the problem and now I'm able to work with the original database.